% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source1.R
\name{generate_data_re2}
\alias{generate_data_re2}
\title{Simulate summary statistics according to the Han and Eskin's RE2 model.  (Heterogeneity only when the population mean effect is non-zero)}
\usage{
generate_data_re2(M = 50 * 10^3, p = 0.01, tau2 = 0.00025,
  resref = NULL, n = rep(50 * 10^3, 10), I2 = 0.3)
}
\arguments{
\item{M}{number of SNPs for which to simulate summary stats.}

\item{p}{proportion of SNPs with real non-zero effects.}

\item{tau2}{variance of real associated non-zero SNPs.}

\item{resref}{an optional vector of residual variances to sample from (with replacement) 
when generating the standard errors for the summary stats.  Default is NULL, in which case the residual variances calculated from a Cigarretes Per Day GWAS (Liu, Jiang 2019)  are used.}

\item{n}{vector of sample sizes from the contributing studies.}

\item{lambda}{proportion of NON-outlier studies for non-replicable SNPs.}

\item{alpha}{variance inflation factor of outlier studies.}
}
\value{
A list containing:
 An Mxk matrix of effect size estimates \code{betajk}, 
 An Mxk matrix of effect size estimate  variances \code{sjk2}, 
 M-length vector inverse-variance weighted meta-analysis z-scores \code{meta.z}, 
 an M-length binary vector indicating real / non-real effect at each SNP \code{Rj},
 an M-length binary vector indicating true effect-size at each SNP \code{muj}.  
 an Mxk matrix of the true study-level effects \code{etajk}
 an M-length vector of the variance of the study-level effects around the SNP's population level effect \code{omega2j}
}
\description{
Simulate summary statistics according to the Han and Eskin's RE2 model.  (Heterogeneity only when the population mean effect is non-zero)
}
\examples{
generate_data_re2(M=100, p=0.01, tau2=2.5e-4, n=rep(10^4, 10), I2=0.2) 
}
