% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source1.R
\name{generateData_fe}
\alias{generateData_fe}
\title{Simulate summary statistics according to the (inverse-variance weighted) fixed effects model.}
\usage{
generateData_fe(M = 50 * 10^3, p = 0.01, tau2 = 0.00025,
  resref = NULL, n = rep(50 * 10^3, 10))
}
\arguments{
\item{M}{number of SNPs for which to simulate summary stats.}

\item{p}{proportion of SNPs with real non-zero effects.}

\item{tau2}{variance of real associated non-zero SNPs.}

\item{resref}{an optional vector of residual variances to sample from (with replacement) 
when generating the standard errors for the summary stats.  Default is NULL, in which case the residual variances calculated from a Cigarretes Per Day GWAS (Liu, Jiang 2019)  are used.}

\item{n}{vector of sample sizes from the contributing studies.}
}
\value{
A list containing:
 An Mxk matrix of effect size estimates \code{betajk}, 
 An Mxk matrix of effect size estimate  variances \code{sjk2}, 
 M-length vector inverse-variance weighted meta-analysis z-scores \code{meta.z}, 
 an M-length binary vector indicating real / non-real effect at each SNP \code{Rj},
 an M-length binary vector indicating true effect-size at each SNP \code{muj}.
}
\description{
Simulate summary statistics according to the (inverse-variance weighted) fixed effects model.
}
\examples{
generateData_fe(M=100, p=0.01, tau2=2.5e-4, resref=NULL, n=rep(50*10^3,10)) 
}
