% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mamba_summary_files.R
\name{make_mamba_summary_files}
\alias{make_mamba_summary_files}
\title{Create input datasets for mamba() function and clumping
 
Take gwas summary stat files, 
  do ivw-meta-analysis, 
  create an assoc dataset which could be used for clumping association statistics in plink,
  create beta and s2 input datasets for mamba() model}
\usage{
make_mamba_summary_files(summary_files,
  standardize_beta_and_s2_datasets = TRUE, beta_name = "b",
  s2_name = "s2", n_name = "n", af_name = "af", ref_name = "ref",
  alt_name = "alt", chr_name = "chr", bp_name = "bp")
}
\arguments{
\item{summary_files}{A list of summary statistic files with common column names}

\item{standardize_beta_and_s2_datasets}{in the outputted b and s2 data tables, 
do you want the effect size estimates and variances standardized (i.e. beta = beta_k*sqrt(2*af_k*(1-af_k)))?  Default is TRUE.}

\item{beta_name}{name of effect size estimate column}

\item{s2_name}{name of effect size variance estimate column}

\item{n_name}{name of n column}

\item{af_name}{name of af column}

\item{ref_name}{name of reference allele column}

\item{alt_name}{name of alt allele column}

\item{chr_name}{name of chr column}

\item{bp_name}{name of bp column}
}
\value{
a list of three data.tables (beta, s2, and assoc)
}
\description{
Create input datasets for mamba() function and clumping
 
Take gwas summary stat files, 
  do ivw-meta-analysis, 
  create an assoc dataset which could be used for clumping association statistics in plink,
  create beta and s2 input datasets for mamba() model
}
\examples{
 \dontrun{
    datadir<-system.file("extdata", package="mamba")
    datafiles<-paste0(datadir, "/", list.files(datadir))
    looky<-make_mamba_summary_files(summary_files=datafiles, 
                             standardize_beta_and_s2_datasets=TRUE,
                             beta_name="betaj", s2_name="sj2")
    looky$assoc[1:5]
    looky$b[1:5] 
    looky$s2[1:5]
}
}
