% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source1.R
\name{mamba}
\alias{mamba}
\title{Fit the MAMBA model.}
\usage{
mamba(betajk, sjk2, parcores = 1, p = 0.003, lambda = 0.96,
  tau2 = 2e-04, alpha = 3, rel.eps = 1e-08, verbose = 1,
  snpids = NA, maxIter = 10^4L)
}
\arguments{
\item{betajk}{Mxk matrix of effect size estimates, where row j corresponds to j-th SNP, and column k corresponds to k-th study .  Missing values can be represented with NA.}

\item{sjk2}{Mxk matrix of effect size estimate variances, where row j corresponds to SNP j, and column k corresponds to study k.  Missing values can be represented with NA.}

\item{p}{initial value for EM algorithm, for the proportion of non-zero SNPs}

\item{lambda}{initial value for EM algorithm, for the proportion of non-replicable SNPs which are well behaved, or non-outliers.}

\item{tau2}{initial value for EM algorithm, for the variance of replicable non-zero effect SNPs.}

\item{alpha}{initial value for EM algorithm, for the variance inflation of outlier summary statistics at non-replicable SNPs.}

\item{rel.eps}{threshold for when to end EM algorithm. rel.eps = (ll[i] - ll[i-1])/ll[i], where ll[i] is the log-likelihood at iteration i.}

\item{snpids}{an optional vector of SNP id names.  If not provided, the ID's will be 1:M, corresponding to the order of the matrix betajk.}

\item{maxIter}{maximum # of EM iterations.}
}
\description{
Fit the MAMBA model.
}
\examples{
  d<-generate_data_mamba()
  mod<-mamba(betajk=d$betajk, sjk2=d$sjk2)
}
