% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mamba_loci.R
\name{select_mamba_loci}
\alias{select_mamba_loci}
\title{Select loci to include in mamba() model, combining clumped and randomly pruned variants.}
\usage{
select_mamba_loci(meta_file, clump_file, bpdiff = 5 * 10^5L,
  prunedsnps_file = NULL, pthresh = NULL)
}
\arguments{
\item{meta_file}{meta-analysis association file (with columns as outputted by mamba::make_mamba_summary_files()$assoc. i.e. plink assoc file format )}

\item{clump_file}{.clumped file output from plink identifying sentinel variants of interest for mamba model}

\item{bpdiff}{BP distance between a randomly pruned snp and a clumped snp of interest.}

\item{prunedsnps_file}{a file with a set of SNP-ids (1 per line) which are approximately independent 
according to a reference panel.  If NULL, a set of SNP id's (in chr:bp_ref_alt format) from HRC reference panel are used.  These were created using plink with parameters --indep-pairwise 500kb 1 0.1,  --maf 0.01}

\item{pthresh}{maximum p-value for index snps of interest.  If not provided, this is taken as maximum p-value of the clumped sentinel variants.}
}
\value{
a data.table with the inputted meta_file, subset to SNPs which are intended to be included in the mamba model.  
 The columns index_snp and prunesnp are indicators for whether the SNP belonged to either the clumped_file or is a randomly pruned marker.
}
\description{
Given 
 1. a set of clumped significant SNPs from meta-analysis which we want to assess with MAMBA, and 
 2. a random set of pruned markers in approximate LD, 
Identify pruned variants no closer than (bpdiff) away from clumped variants, 
to be used in fitting mamba, to be used in fitting mamba() model.  
 The inclusion of randomly pruned variants allows the non-replicable zero-effect cluster of the mixture model to be reliably estimated.
}
\examples{
 \dontrun{
   ## see online vignette for detailed example. 
    
}
}
